+++ packages/remark-mdx/.babelrc
@@ -1,0 +1,13 @@
+{
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "browsers": "last 2 versions",
+          "node": "8"
+        }
+      }
+    ]
+  ]
+}
