+++ packages/remark-mdx/package.json
@@ -2,6 +2,7 @@
   "name": "remark-mdx",
   "version": "1.0.2",
   "description": "Support import, export, and JSX in markdown",
+  "main": "dist/index.js",
   "license": "MIT",
   "keywords": [
     "mdx",
@@ -22,29 +23,36 @@
     "Titus Wormer <tituswormer@gmail.com> (https://wooorm.com)"
   ],
   "files": [
-    "index.js",
-    "tag.js",
-    "block.js",
-    "extract-imports-and-exports.js"
+    "dist",
+    "license",
+    "readme.md"
   ],
   "dependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/helper-plugin-utils": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
-    "@babel/plugin-syntax-jsx": "^7.2.0",
     "is-alphabetical": "^1.0.2",
     "remark-parse": "^6.0.0",
     "unified": "^7.0.0"
   },
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "build": "babel src --out-dir dist --env-name production --config-file ./.babelrc",
+    "start": "babel src --out-dir dist --watch --env-name development --config-file  ./.babelrc"
   },
   "jest": {
     "testEnvironment": "node"
   },
+  "optionalDependencies": {
+    "@babel/helper-plugin-utils": "^7.0.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.3.2",
+    "@babel/plugin-syntax-jsx": "^7.2.0",
+    "@babel/core": "^7.2.2",
+    "@babel/standalone": "^7.2.2"
+  },
   "devDependencies": {
+    "@babel/cli": "^7.2.2",
     "jest": "^24.0.0",
     "remark-stringify": "^6.0.4",
-    "vfile": "^4.0.0"
+    "vfile": "^4.0.0",
+    "@babel/core": "^7.2.2",
+    "@babel/standalone": "^7.2.2"
   }
 }
