+++ packages/remark-mdx/test/import-export.test.js
@@ -1,4 +1,4 @@
-const extract = require('../extract-imports-and-exports')
+const extract = require('../dist/extract-imports-and-exports')
 const fixtures = require('./fixtures/import-export')
 const vfile = require('vfile')
