+++ packages/remark-mdx/test/inline-parsing.test.js
@@ -1,7 +1,7 @@
 const unified = require('unified')
 const remarkParse = require('remark-parse')
 const remarkStringify = require('remark-stringify')
-const remarkMdx = require('..')
+const remarkMdx = require('../dist')
 
 const fixtures = require('./fixtures/inline-parsing')
